requirements.txt
asgiref==3.8.1
atomicwrites==1.4.1
attrs==22.1.0
colorama==0.4.6
Django==4.2.16
djangorestframework==3.15.2
iniconfig==1.1.1
mypy==1.2.0
mypy-extensions==1.0.0
packaging==21.3
pillow==10.4.0
pluggy==0.13.1
py==1.11.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2022.6
sqlparse==0.5.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2024.2
